cleans up Ang page, crates multiple hooks
