+++ src/js/hooks/use-key-down-event-handler.ts
@@ -1,0 +1,9 @@
+import { useEffect } from 'react';
+
+export const useKeydownEventHandler = (eventListener: EventListener) => {
+  useEffect(() => {
+    document.addEventListener('keydown', eventListener);
+
+    return () => document.removeEventListener('keydown', eventListener);
+  }, [eventListener])
+}
