+++ src/js/pages/Ang/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './change-ang-in-view';
+export * from './change-highlighted-pankti';
